requirements.txt
fastapi
torch
torchvision
pillow
python-multipart
uvicorn[standard]
